+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.vue
@@ -1,10 +1,15 @@
 <script lang="ts" setup>
+import { OnyxLoadingIndicator } from "@/index";
 import { computed } from "vue";
 import type { OnyxInputProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxInputProps>(), {
   modelValue: "",
   type: "text",
+  autocapitalize: "sentences",
+  readonly: false,
+  disabled: false,
+  loading: false,
 });
 
 const emit = defineEmits<{
@@ -38,6 +43,11 @@ const handleChange = (event: Event) => {
   const inputValue = (event.target as HTMLInputElement).value;
   emit("change", inputValue);
 };
+
+const patternSource = computed(() => {
+  if (props.pattern instanceof RegExp) return props.pattern.source;
+  return props.pattern;
+});
 </script>
 
 <template>
@@ -47,22 +57,36 @@ const handleChange = (event: Event) => {
     </span>
 
     <div class="onyx-input__wrapper">
+      <OnyxLoadingIndicator v-if="props.loading" class="onyx-input__loading" type="circle" />
+
+      <!-- eslint-disable vuejs-accessibility/no-autofocus -
+         We want to provide the flexibility to have the autofocus property.
+         The JSDoc description includes a warning that it should be used carefully.
+      -->
       <input
         v-model="value"
         class="onyx-input__native"
         :placeholder="props.placeholder"
         :type="props.type"
+        :autocapitalize="props.autocapitalize"
+        :autocomplete="props.autocomplete"
+        :autofocus="props.autofocus"
+        :name="props.name"
+        :pattern="patternSource"
+        :readonly="props.readonly"
+        :disabled="props.disabled || props.loading"
         @change="handleChange"
         @focus="emit('focus')"
         @blur="emit('blur')"
       />
+      <!-- eslint-enable vuejs-accessibility/no-autofocus -->
     </div>
   </label>
 </template>
 
 <style lang="scss">
 .onyx-input {
-  --border-color: var(--onyx-color-base-primary-500);
+  --border-color: var(--onyx-color-base-neutral-300);
   --selection-color: var(--onyx-color-base-primary-200);
 
   font-family: var(--onyx-font-family);
@@ -79,8 +103,8 @@ const handleChange = (event: Event) => {
 
   &__wrapper {
     border-radius: var(--onyx-radius-sm);
-    border: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-300);
-    background: var(--onyx-color-base-background-blank);
+    border: var(--onyx-1px-in-rem) solid var(--border-color);
+    background-color: var(--onyx-color-base-background-blank);
     color: var(--onyx-color-text-icons-neutral-intense);
 
     display: flex;
@@ -90,22 +114,39 @@ const handleChange = (event: Event) => {
     font-size: 1rem;
     line-height: $line-height;
 
-    height: calc($line-height + 2 * $padding-vertical);
     box-sizing: border-box;
+    padding: $padding-vertical var(--onyx-spacing-sm);
+    height: calc($line-height + 2 * $padding-vertical);
 
-    &:has(.onyx-input__native:enabled:hover) {
+    &:has(.onyx-input__native:read-write:hover) {
       border-color: var(--onyx-color-base-primary-400);
     }
 
     &:has(.onyx-input__native:enabled:focus) {
-      border-color: var(--border-color);
+      --border-color: var(--onyx-color-base-primary-500);
       outline: var(--onyx-spacing-4xs) solid var(--onyx-color-base-primary-200);
+    }
+
+    // :read-only is valid for readonly and disabled state so we put shared styles for both states here
+    &:has(.onyx-input__native:read-only) {
+      --selection-color: var(--onyx-color-base-neutral-200);
+      background-color: var(--onyx-color-base-background-tinted);
+    }
+
+    // styles for readonly but NOT disabled
+    &:has(.onyx-input__native:enabled:read-only) {
+      &:has(.onyx-input__native:hover) {
+        --border-color: var(--onyx-color-base-neutral-400);
+      }
+
+      &:has(.onyx-input__native:focus) {
+        --border-color: var(--onyx-color-base-neutral-500);
+        outline-color: var(--onyx-color-base-neutral-200);
+      }
     }
   }
 
   &__native {
-    padding: $padding-vertical var(--onyx-spacing-sm);
-
     // reset native input styles so they are inherited from the parent
     border: none;
     border-radius: inherit;
@@ -125,6 +166,22 @@ const handleChange = (event: Event) => {
     &::selection {
       background: var(--selection-color);
     }
+  }
+
+  &:has(&__native:disabled) {
+    .onyx-input {
+      &__label {
+        color: var(--onyx-color-text-icons-neutral-soft);
+      }
+
+      &__wrapper {
+        color: var(--onyx-color-text-icons-neutral-soft);
+      }
+    }
+  }
+
+  &__loading {
+    color: var(--onyx-color-text-icons-primary-intense);
   }
 }
 </style>